// src/utils/enums.ts
import { LANG, LANG_OBJ } from "./lang";

export enum SYSTEM_MODE {
  MANUAL = "MANUAL",
  AUTO = "AUTO",
}

export enum WARNING_LEVEL {
  AMBER = "AMBER",
  RED = "RED",
  BLACK = "BLACK",
  CANCEL = "CANCEL",
}

export enum ACTION_CODE {
  ISSUE = "ISSUE",
  CANCEL = "CANCEL",
}

export enum WARNING_CODE {
  HSWW_AMBER_OUTDOOR = "HSWW_AMBER_OUTDOOR",
  HSWW_RED_OUTDOOR = "HSWW_RED_OUTDOOR",
  HSWW_BLACK_OUTDOOR = "HSWW_BLACK_OUTDOOR",
  HSWW_CANCEL_OUTDOOR = "HSWW_CANCEL_OUTDOOR",

  HSWW_AMBER_INDOOR = "HSWW_AMBER_INDOOR",
  HSWW_RED_INDOOR = "HSWW_RED_INDOOR",
  HSWW_BLACK_INDOOR = "HSWW_BLACK_INDOOR",
  HSWW_CANCEL_INDOOR = "HSWW_CANCEL_INDOOR",
}

export enum WARNING_STATUS {
  ISSUE = "ISSUE",
  CANCEL = "CANCEL",
}

export const SYSTEM_MODE_NAME = {
  [SYSTEM_MODE.AUTO]: LANG(LANG_OBJ.SYSTEM_MODE.AUTO),
  [SYSTEM_MODE.MANUAL]: LANG(LANG_OBJ.SYSTEM_MODE.MANUAL),
};

export const HSWW_ICON_MAPPING = {
  [WARNING_LEVEL.AMBER]: "/images/amber.png",
  [WARNING_LEVEL.RED]: "/images/red.png",
  [WARNING_LEVEL.BLACK]: "/images/black.png",
  [WARNING_LEVEL.CANCEL]: "/images/cancel.png",
};

export const HSWW_COLOR_MAPPING = {
  [WARNING_LEVEL.AMBER]: LANG(LANG_OBJ.WARNING_COLOR.AMBER),
  [WARNING_LEVEL.RED]: LANG(LANG_OBJ.WARNING_COLOR.RED),
  [WARNING_LEVEL.BLACK]: LANG(LANG_OBJ.WARNING_COLOR.BLACK),
  [WARNING_LEVEL.CANCEL]: LANG(LANG_OBJ.WARNING_COLOR.CANCEL),
};